// src/app/generated/model/themeUpdateCreate.ts
/**
 * onecx-theme-bff
 * OneCx theme Bff
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ThemeUpdateCreate { 
    version?: number;
    creationDate?: string;
    creationUser?: string;
    modificationDate?: string;
    modificationUser?: string;
    name?: string;
    cssFile?: string;
    description?: string;
    assetsUrl?: string;
    logoUrl?: string;
    faviconUrl?: string;
    previewImageUrl?: string;
    assetsUpdateDate?: string;
    properties?: object;
}