// src/app/generated/model/updateThemeResponse.ts
/**
 * onecx-theme-bff
 * OneCx theme Bff
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Theme } from './theme';


export interface UpdateThemeResponse { 
    resource: Theme;
}